SignUp serializer validates on if usernames exist for phone number
I think that basically this validator would receive the already-updated hash preivously, so the iexact filter was being run against not the phone number, but the mapped username.

So here, we just check to see if there are usernames for the phone number, and if there are, we raise the error, otherwise we return the create_new_username, which both sets the value of username to the hash generated and creates the mapping between the phone number and the username.